+++ Program.cs
@@ -191,174 +191,6 @@ namespace UtilityBelt
 
     #region Choice Processors
 
-    #region Covid-19
-
-    private static void Covid19()
-    {
-      logger.LogInformation($"User Choice : {nameof(Covid19)}");
-      while (true)
-      {
-        Console.WriteLine("");
-        Console.WriteLine("Enter the country name to get the information. If you want to see the global information, type \"Global\". For the list of all countries type \"List\". To exit Covid-19 Statistics type \"Exit\": ");
-        string userInput = Console.ReadLine();
-        Console.WriteLine("");
-
-        if (userInput == "Exit")
-          return;
-
-        string content = string.Empty;
-        using (var wc = new WebClient())
-        {
-          content = wc.DownloadString("https://api.covid19api.com/summary");
-        }
-        CovidRoot summary = JsonSerializer.Deserialize<CovidRoot>(content);
-
-        if (userInput.StartsWith("List", StringComparison.InvariantCultureIgnoreCase))
-        {
-          userInput = userInput.Substring(4).TrimStart();
-          Console.WriteLine("Found countries: ");
-          foreach (var countryJson in summary.Countries)
-          {
-            if (countryJson.Country.StartsWith(userInput, StringComparison.InvariantCultureIgnoreCase))
-              Console.WriteLine(countryJson.Country);
-          }
-        }
-        else if (userInput.Equals("Global", StringComparison.InvariantCultureIgnoreCase))
-        {
-          ShowCovidInfo("Global", summary.Global.NewConfirmed, summary.Global.TotalConfirmed, summary.Global.NewDeaths, summary.Global.TotalDeaths, summary.Global.NewRecovered, summary.Global.TotalRecovered);
-        }
-        else
-        {
-          bool countryExists = false;
-
-          foreach (var countryJson in summary.Countries)
-          {
-            if (userInput.Equals(countryJson.Country, StringComparison.InvariantCultureIgnoreCase))
-            {
-              ShowCovidInfo(countryJson.Country, countryJson.NewConfirmed, countryJson.TotalConfirmed, countryJson.NewDeaths, countryJson.TotalDeaths, countryJson.NewRecovered, countryJson.TotalRecovered);
-              countryExists = true;
-            }
-          }
-
-          if (countryExists == false)
-            Console.WriteLine("Country does not exist. Type \"List\" to see the list of available countries.");
-        }
-      }
-    }
-
-    private static void ShowCovidInfo(string countryName, long newConfirmed, long totalConfirmed, long newDeaths, long totalDeaths, long newRecovered, long totalRecovered)
-    {
-      Console.WriteLine("Statistics: " + countryName);
-      Console.ForegroundColor = ConsoleColor.Cyan;
-      Console.WriteLine("New Confirmed: " + newConfirmed);
-      Console.WriteLine("Total Confirmed: " + totalConfirmed);
-      Console.ForegroundColor = ConsoleColor.Red;
-      Console.WriteLine("New Deaths: " + newDeaths);
-      Console.WriteLine("Total Deaths: " + totalDeaths);
-      Console.ForegroundColor = ConsoleColor.Green;
-      Console.WriteLine("New Recovered: " + newRecovered);
-      Console.WriteLine("Total Recovered: " + totalRecovered);
-    }
-
-    #endregion Covid-19
-
-    #region GeekJokes
-
-    private static void GeekJokes()
-    {
-      logger.LogInformation($"User Choice : {nameof(GeekJokes)}");
-      IWebProxy defaultWebProxy = WebRequest.DefaultWebProxy;
-      defaultWebProxy.Credentials = CredentialCache.DefaultCredentials;
-
-      string content = string.Empty;
-      string geekJokeUrl = "https://geek-jokes.sameerkumar.website/api?format=json";
-      using (var wc = new WebClient() { Proxy = defaultWebProxy })
-      {
-        content = wc.DownloadString(geekJokeUrl);
-      }
-
-      GeekJokeModel joke = JsonSerializer.Deserialize<GeekJokeModel>(content);
-      Console.WriteLine();
-      Console.WriteLine(@$"The Geek says -- {joke.Joke}");
-
-      Console.WriteLine();
-    }
-
-    #endregion GeekJokes
-
-    #region NumberFact
-
-    private static void NumberFact()
-    {
-      logger.LogInformation($"User Choice : {nameof(NumberFact)}");
-      string content = string.Empty;
-      int numberEntered;
-
-      Console.WriteLine("Please enter an integer number");
-      String userInput = Console.ReadLine();
-
-      if (int.TryParse(userInput, out numberEntered))
-      {
-        string numberFactURL = $"http://numbersapi.com/{userInput}?format=json";
-        using (var wc = new WebClient())
-        {
-          content = wc.DownloadString(numberFactURL);
-        }
-
-        Console.WriteLine("Random fact for number " + userInput + ":");
-        Console.WriteLine(content);
-      }
-      else
-      {
-        Console.WriteLine("Input was not an integer");
-      }
-    }
-
-    #endregion NumberFact
-
-    #region Random Advice
-
-    private static void RandomAdvice()
-    {
-      logger.LogInformation($"User Choice : {nameof(RandomAdvice)}");
-      string content = string.Empty;
-      string adviceUrl = "https://api.adviceslip.com/advice";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(adviceUrl);
-      }
-
-      AdviceModel randomAdvice = JsonSerializer.Deserialize<AdviceModel>(content);
-      Console.WriteLine("Here's your advice:");
-      Console.ForegroundColor = ConsoleColor.Cyan;
-      Console.WriteLine(randomAdvice.Slip.Advice);
-      Console.WriteLine();
-    }
-
-    #endregion Random Advice
-
-    #region Space Station Location
-
-    private static void SpaceStationLocation()
-    {
-      logger.LogInformation($"User Choice : {nameof(SpaceStationLocation)}");
-      string content = string.Empty;
-      string adviceUrl = "http://api.open-notify.org/iss-now.json";
-      using (var wc = new WebClient())
-      {
-        content = wc.DownloadString(adviceUrl);
-      }
-
-      SpaceStationLocation spaceStation = JsonSerializer.Deserialize<SpaceStationLocation>(content);
-      Console.WriteLine("Woah! The International Space Station is currently at:");
-      Console.ForegroundColor = ConsoleColor.Cyan;
-      Console.WriteLine($@"{spaceStation.Location.Latitude} Latitude and {spaceStation.Location.Longitude} Longitude");
-      Console.Write("That's so cool!");
-      Console.WriteLine();
-    }
-
-    #endregion Space Station Location
-
     #region Random quote
 
     private static void RandomQuoteGarden()
